Fix bug making all tests fail when including optional = false (WTF?!)
I really don't understand what happened here.

Every single test started failing as though I was violating a not null constraint in the DB as soon as I added `optional = false` to the `ManyToOne` relationship between concrete and info classes, for example, `Book` and `BookInfo`.

One of the errors in the (god awful) stacktrace Spring gave me when a test was failing included something along the lines of "`album_info_id` column should not be null" when I was debugging a test specifically about the `Book` class.

Therefore, I hypothesized that for some (probably just as god awful) reason, Spring added that not null constraint to every column that held one of those non-optional foreign key constraints in the single table that holds all LibraryItem, effectively forcing you to link every single `Book` instance to hold an association to a `BookInfo` instance, but also an `AlbumInfo` instance, a `MovieInfo` instance, etc.

That hypothesis seems to have confirmed itself with this (hacky) solution: changing the inheritance strategy to `JOINED` or `TABLE_PER_CLASS` makes all the tests pass again. Anything else makes all the tests fail.